Implemented Login and Logout Functionality
