+++ src/main/java/com/libraryman_api/security/controllers/SignupController.java
@@ -1,4 +1,4 @@
-package com.libraryman_api.security;
+package com.libraryman_api.security.controllers;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.libraryman_api.member.Members;
+import com.libraryman_api.security.services.SignupService;
 
 
 @RestController
@@ -21,5 +22,13 @@ public class SignupController {
 	public void signup(@RequestBody Members members) {
 		this.signupService.signup(members);
 		
+	}
+	@PostMapping("/api/signup/admin")
+	public void signupAdmin(@RequestBody Members members) {
+		this.signupService.signupAdmin(members);
+	}
+	@PostMapping("/api/signup/librarian")
+	public void signupLibrarian(@RequestBody Members members) {
+		this.signupService.signupLibrarian(members);
 	}
 }
