+++ src/main/java/com/libraryman_api/security/jwt/JwtAuthenticationFilter.java
@@ -1,4 +1,4 @@
-package com.libraryman_api.security;
+package com.libraryman_api.security.jwt;
 
 
 import java.io.IOException;
