+++ src/main/java/com/libraryman_api/security/model/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.libraryman_api.security;
+package com.libraryman_api.security.model;
 
 public class LoginRequest {
