+++ src/main/java/com/libraryman_api/security/model/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.libraryman_api.security;
+package com.libraryman_api.security.model;
 
 
 public class LoginResponse {
