+++ src/main/resources/application.properties
@@ -1,2 +1,3 @@
 spring.application.name=libraryman-api
-spring.profiles.active=${ENV:dev}
+spring.profiles.active=${ENV:production}
+jwt.secretKey=${YOUR_JWT_SECRET_KEY}
